Initial Change Data Capture Work
The goal of these commits is to capture changes as they are written
to LevelDB so they can be used to replicate the database to another
server.

The work so far defines a database wrapper, exposing the
ethdb.Database interface and wrapping both another ethdb.Database
and a cdc.LogProducer. Any operations made to the wrapper database
get proxied to the wrapped database, and the transactions are
relayed to the LogProducer.

For now we have a MockLogProducer and MockLogConsumer for testing,
and a KafkaLogProducer and KafkaLogConsumer for operational use.

The next steps will be to add options to launch a node configured
with the KafkaLogProducer, which will write database operations out
to Kafka, and a node that uses the KafkaLogConsumer to populate the
database from the Kafka without connecting to other p2p nodes.