Add kafka options to the geth command
This adds flags to start a geth node that will send database writes
to kafka, or write to the database from logs that come from Kafka.

The current implementation kind of works. The read replicas won't recognize
when block updates get recorded, so they continue to report stale data
for the "latest" block, but if asked for information from specific,
more recent blocks it will return the correct information.