Consolidate CDC flags
Normal Geth nodes will only ever publish, and replica nodes will
only ever subscribe, so we don't need the flags to specify source
and destination brokers and topics.

Originally the idea was that a replica was just going to be a
normal geth node with the source flag, but now that we've built
out a separate replica command that no longer applies.