Generic mechanism for database upgrades.
The database upgrade system assumed that the database would always
be a LevelDB instance, and was crashing on cdc.DBWrappers. This
introduces an IterableDatabase interface, which can apply to both
the LevelDB database and the wrapper.

It still returns a leveldb iterator, which I'm not thrilled with,
but it's more robust than just assuming the object will always be
a LevelDB object.