Don't start node until synced with partition
We were getting some errors if the replica tried to start serving
before it had synced with the partition. It also makes sense that
we don't want it serving over RPC before it has pulled the latest
state.

I'm still seeing some issues, but those seem to be in the replica
code at this point.