Initial Commit for working writable slave replicas
This is a decent overhaul, where I used the changes made for
KafkaLogTopic as a base.

Created similarly named KafkaTransactionTopic, flags, and etc.

Changed the geth/chaincmd.go to use correct chainid due to
chainid not being a public function on this version of Geth

Created syncronous kafka producer, instead of an async
since users would likely enjoy knowing their transaction sending failed.

Extended the documentation so it can be repeated by others