ethdb.cdc: Properly handle batch deletes
When I originally implemented deletes in batches, I didn't perform
the delete as part of the atomic batch operation, and instead just
did the deletes out of band. This refactors batch to incorporate
deletes properly.

I also fixed some CDc related tests that weren't passing due to other
changes.

This is an intermediate step. Next I'm going to make it so that batch
operations get sent across the wire sequentially, rather than as
a single message that may be too large for Kafka to handle.