ethdb/cdc, replica: Log replica status periodically
This logs several pieces of information on a regular basis to help
with monitoring replicas. Logged data includes:

* Block Number
* Block Hash
* Block Age
* Topic Offset
* Topic Offset Age

Reporting the topic offset and topic offset age required being able
to look up that information from leveldb, so the topic name had to be
made available to the replica, hence the changes in the cdc module.