cmd/geth: Add sethead subcommand
This adds a sethead subcommand, which allows us to roll back to an
earlier block. When snapshotting replicas, it's possible that a replica
will be in the middle of a master's write operation when it shuts
down or is snapshotted. This allows us to roll back to an earlier block
that should have been written completely before using that snapshot
to start a master.