replica, ethdb/cdc, cmd/geth: Add replication heartbeats
This has the master publish periodic heartbeats to Kafka. This will
allow replicas to identify when the master has ceased communication,
as it can expect some communication on a regular interval.

Additionally, we added a --replica.startup.age flag, which indicates
the maximum number of seconds old the head block may be for the replica
to start serving. This is to help address the scenario where we are
standing up a new cluster behind an existing load balancer - the replica
will not start serving HTTP (and thus will not join the load balancer)
until it has received a recent block. If this flag is not set, there is
no age limit, and replicas will begin serving as soon as they are in
sync with the master.