ethdb/overlay, ethdb/devnull, cmd/geth: Add overlay and null storage backends
Adds an overlay database wrapper, which combines an overlay and underlay
database. Because EBS snapshots are subject to cold starts, we're able
to get much faster starts by making the snapshot an underlay and
writing everything to kafka on an NVME or blank EBS volume.

It also adds a null backend, which responds affirmatively to writes,
but doesn't track anything. This is used mainly for testing.

In the replica command line, the overlay can be specified as a path,
in which case it will be initialized as leveldb if none is there.
Altenratively, the overlay can be specified as "null" or "mem" to
use the devnull backend or memorydb respectively.