replica, ethdb/cdc: Enable websockets and graphql
This commit adds websockets and graphql to the replica service.

Graphql was pretty simple, but required a bit of restructuring how
the replica service gets set up in the first place so that graphql
can be started along with it.

Websocket subscriptions required being able to identify new blocks,
new logs, and reverted logs and publish them through feeds. This
was accomplished by hooking into our watch for the LastBlock key.
When the LastBlock key is written, we retrieve the related block,
find any blocks that got uncled as a result, and publish the logs
of both the uncled blocks (as reverted) and the new blocks.