replica: Better transaction validation
Prior to this, we were just passing transactions to kafka and letting
the txrelay worry about validity. The lack of error messages is causing
problems for some users, so we're adding the same validation that the
txpool does. In the future, we may be able to leverage the txpool
more directly.