replica, ethdb/cdc, replicacmd, txrelay: Replica support for mempool transactions
Feed transactions from Kafka into the mempool.

Also modifies CreateTopicIfDoesNotExist to accept a number of partitions.
For the CDC stream we've alway used a single partition to ensure in-order
delivery and simplify the tracking of offsets. With mempool data,
order is not important, and since we'll only start pulling the latest
data when a replica starts up, we don't need to track offsets.