rpc: Better context cancellation
This commit propagates context from various RPC providers to
the message handler. Prior to this, if you cancelled an HTTP
request it would cease executing the request, but websockets
and IPC requests would continue executing. This ensures that
the websockets request context is used for request execution,
and creates a context completes with the IPC connection closes.