replica: Reprocess events on startup
When a master starts we want to replay the events from the last ten
blocks, just to make sure they made it into the feed. We've seen
some instances of masters shutting down after recording the latest
blocks into leveldb (and the corresponding Kafka topic), but not
emitting events first. This leads to a scenario where the final
block before shutdown never makes it into the feed.

One block here is probably sufficient, but having blocks appear in
the feed multiple times is safer than not at all, so we're being
very cautious by emitting ten blocks into the stream.